Spawnowanie nowych kółek
